+++ src/main/java/org/opensearch/neuralsearch/ml/MLCommonsClientAccessor.java
@@ -636,6 +636,7 @@ public class MLCommonsClientAccessor {
      * @param agenticQuery agentic query
      * @param agentId agent id
      * @param agentInfo agent info
+     * @param embeddingModelId embedding model id
      * @param xContentRegistry xContentRegistry
      * @param listener listener to be called with the agent execution result
      */
@@ -644,10 +645,11 @@ public class MLCommonsClientAccessor {
         @NonNull AgenticSearchQueryBuilder agenticQuery,
         @NonNull String agentId,
         @NonNull AgentInfoDTO agentInfo,
+        String embeddingModelId,
         @NonNull NamedXContentRegistry xContentRegistry,
         @NonNull ActionListener<AgentExecutionDTO> listener
     ) throws IOException {
-        retryableExecuteAgent(request, agenticQuery, agentId, agentInfo, xContentRegistry, 0, listener);
+        retryableExecuteAgent(request, agenticQuery, agentId, agentInfo, embeddingModelId, xContentRegistry, 0, listener);
     }
 
     /**
@@ -658,6 +660,7 @@ public class MLCommonsClientAccessor {
         AgenticSearchQueryBuilder agenticQuery,
         String agentId,
         AgentInfoDTO agentInfo,
+        String embeddingModelId,
         NamedXContentRegistry xContentRegistry,
         int retryTime,
         ActionListener<AgentExecutionDTO> listener
@@ -700,6 +703,10 @@ public class MLCommonsClientAccessor {
             parameters.put("query_fields", gson.toJson(agenticQuery.getQueryFields()));
         }
 
+        if (embeddingModelId != null && !embeddingModelId.trim().isEmpty()) {
+            parameters.put("embedding_model_id", embeddingModelId);
+        }
+
         if (hasSystemPrompt == false && type != MLAgentType.FLOW) {
             parameters.put("system_prompt", loadSystemPrompt());
         }
@@ -740,7 +747,16 @@ public class MLCommonsClientAccessor {
             listener.onResponse(new AgentExecutionDTO(removeTrailingDecimalZeros(dslQuery), agentStepsSummary, memoryId, selectedIndex));
         }, e -> RetryUtil.handleRetryOrFailure(e, retryTime, () -> {
             try {
-                retryableExecuteAgent(request, agenticQuery, agentId, agentInfo, xContentRegistry, retryTime + 1, listener);
+                retryableExecuteAgent(
+                    request,
+                    agenticQuery,
+                    agentId,
+                    agentInfo,
+                    embeddingModelId,
+                    xContentRegistry,
+                    retryTime + 1,
+                    listener
+                );
             } catch (IOException ex) {
                 throw new RuntimeException(ex);
             }